+++ src/components/search/algolia/AlgoliaSearchDashboard.tsx
@@ -37,7 +37,7 @@ const AlgoliaSearchContent: React.FC = () => {
           {/* Results Section */}
           <section className="lg:col-span-3">
             <Configure 
-              hitsPerPage={36} 
+              hitsPerPage={36}
               ruleContexts={[`origin:${origin}`]}
             />
             <SearchStats />